+++ library/src/test/java/cloud/pace/sdk/appkit/IconUtilsTest.kt
@@ -3,40 +3,60 @@ package cloud.pace.sdk.appkit
 import cloud.pace.sdk.appkit.model.AppManifest
 import cloud.pace.sdk.utils.IconUtils
 import junit.framework.Assert.assertEquals
+import junit.framework.Assert.assertNull
 import org.junit.Test
 
 class IconUtilsTest {
 
-    @Test
-    fun `get png icon`() {
-        val buttonWidth = 80.0
-        val pngIcon = AppManifest.AppIcons("logo.png", "12x12 50x50 120x120", "image/png")
-        val favIcon = AppManifest.AppIcons("favicon.ico", "64x64 32x32 24x24 16x16", "image/x-icon")
-
-        val bestIcon = IconUtils.getBestMatchingIcon(buttonWidth, arrayOf(pngIcon, favIcon))
-        assertEquals(pngIcon.src, bestIcon?.src)
-        assertEquals(pngIcon.type, bestIcon?.type)
-    }
+    private val requestedSize = 64.0
 
     @Test
-    fun `get best png icon`() {
-        val buttonWidth = 80.0
+    fun `get best png icon from multiple icons`() {
         val smallPngIcon = AppManifest.AppIcons("smallLogo.png", "12x12 30x30", "image/png")
-        val pngIcon = AppManifest.AppIcons("logo.png", "50x50 120x120", "image/png")
+        val appIcon1 = AppManifest.AppIcons("logo1.png", "50x50 120x120", "image/png")
+        val appIcon2 = AppManifest.AppIcons("logo2.png", "12x12 16x21 65x63", "image/png")
+        val appIcon3 = AppManifest.AppIcons("logo3.png", "120x120 128x64", "image/png")
+        val appIcon4 = AppManifest.AppIcons("logo4.png", "45x67 32x32", "image/png")
+        val appIcon5 = AppManifest.AppIcons("logo5.png", "asdf", "image/png")
         val favIcon = AppManifest.AppIcons("favicon.ico", "64x64 32x32 24x24 16x16", "image/x-icon")
 
-        val bestIcon = IconUtils.getBestMatchingIcon(buttonWidth, arrayOf(smallPngIcon, pngIcon, favIcon))
-        assertEquals(pngIcon.src, bestIcon?.src)
-        assertEquals(pngIcon.type, bestIcon?.type)
+        val bestIcon = IconUtils.getBestMatchingIcon(requestedSize, arrayOf(smallPngIcon, appIcon1, appIcon2, appIcon3, appIcon4, appIcon5, favIcon))
+        assertEquals(appIcon2, bestIcon)
     }
 
     @Test
     fun `only favicon available`() {
-        val buttonWidth = 80.0
         val favIcon = AppManifest.AppIcons("favicon.ico", "64x64 32x32 24x24 16x16", "image/x-icon")
 
-        val bestIcon = IconUtils.getBestMatchingIcon(buttonWidth, arrayOf(favIcon))
-        assertEquals(favIcon.src, bestIcon?.src)
-        assertEquals(favIcon.type, bestIcon?.type)
+        val bestIcon = IconUtils.getBestMatchingIcon(requestedSize, arrayOf(favIcon))
+        assertEquals(favIcon, bestIcon)
+    }
+
+    @Test
+    fun `only one png icon available`() {
+        val appIcon = AppManifest.AppIcons("logo.png", "32x32", "image/png")
+        val selectedIcon = IconUtils.getBestMatchingIcon(requestedSize, arrayOf(appIcon))
+        assertEquals(appIcon, selectedIcon)
+    }
+
+    @Test
+    fun `no icon available`() {
+        val selectedIcon = IconUtils.getBestMatchingIcon(requestedSize, emptyArray())
+        assertNull(selectedIcon)
+    }
+
+    @Test
+    fun `no icon available because of missing sizes for one icon`() {
+        val appIcon = AppManifest.AppIcons("logo.png", "", "image/png")
+        val selectedIcon = IconUtils.getBestMatchingIcon(requestedSize, arrayOf(appIcon))
+        assertNull(selectedIcon)
+    }
+
+    @Test
+    fun `no icon available because of missing sizes for all icons`() {
+        val appIcon1 = AppManifest.AppIcons("logo1.png", "", "image/png")
+        val appIcon2 = AppManifest.AppIcons("logo2.png", "", "image/png")
+        val selectedIcon = IconUtils.getBestMatchingIcon(requestedSize, arrayOf(appIcon1, appIcon2))
+        assertNull(selectedIcon)
     }
 }